requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.17
websockets==13.1
kafka-python==2.0.2
pandas==2.2.3
pytest==8.3.3
httpx==0.27.2
